api/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
dotenv==0.9.9
Markdown==3.7
mysqlclient==2.2.7
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
